Move bintray setup in library module
