--- buildSrc/src/main/kotlin/Publish.kt
@@ -1,15 +1,0 @@
-/**
- * See:
- * https://github.com/codepath/android_guides/wiki/Building-your-own-Android-library
- *
- * For gradle config:
- * https://inthecheesefactory.com/blog/how-to-upload-library-to-jcenter-maven-central-as-dependency/en
- */
-object Publish {
-    const val groupId = "com.github.pgreze"
-    const val artifactId = "android-reactions"
-    const val version = "1.0"
-
-    const val siteUrl = "https://github.com/pgreze/android-reactions"
-    const val gitUrl = "$siteUrl.git"
-}
